long -> double

+++ тркпо/60f_big_int/src/lab_big_int/AvgLongClassTest.java
@@ -29,13 +29,13 @@ class AvgLongClassTest {
     void averageLongTest3() {
         ArrayList<Long> numbers = Stream.of(9_000_000_000_000_000_000L, 8_444_444_444_444_444_444L)
                 .collect(Collectors.toCollection(ArrayList::new));
-        assertEquals(9_000_000_000_000_000_000L/2 + 8_444_444_444_444_444_444L/2, AvgLongClass.averageLong(numbers));
+        assertEquals((double) 9_000_000_000_000_000_000L/2 + (double) 8_444_444_444_444_444_444L/2, AvgLongClass.averageLong(numbers));
     }
 
     @Test
     void averageLongTest4() {
         ArrayList<Long> numbers = Stream.of(Long.MAX_VALUE, Long.MIN_VALUE)
                 .collect(Collectors.toCollection(ArrayList::new));
-        assertEquals(Long.MAX_VALUE/2 + Long.MIN_VALUE/2, AvgLongClass.averageLong(numbers));
+        assertEquals((double) Long.MAX_VALUE/2 + (double) Long.MIN_VALUE/2, AvgLongClass.averageLong(numbers));
     }
 }
